More general example script
